Improve component API design
* Renamed QuestionCards to QuizNavigator and QuestionCard to
AnswerPicker to align the component names with their function.

*Changed QuizNavigator's API from (data, onAnswerSelect, quizSubmitted)
to (questionsData, onAnswerPick, disableNav) because disableNav is more
suitable than quizSubmitted given then function of the component.
Navigator need not know whether the quiz is submitted or not, it only
needs to know whether it's disabled or enabled.

* Changed AnswerPicker's API from (question, choices, onAnswerSelect) to
(questionIndex, question, answers, onAnswerPick). This make the API
consistent with other components by using the same word "answer"
throughout. It also moves the picking logic to the Picker, which was
initially in the Navigator.

* Renamed QuizProgress to DiscreteProgress because it didn't have
business logic and it is highly reusable. Now it is clear what spot it
occupied on the spectrum of components